requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cairo=1.16.0=h19f5f5c_2
certifi=2020.6.20=pyhd3eb1b0_3
charset-normalizer=2.0.12=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
dominate=2.4.0=pypi_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
gdk-pixbuf=2.42.8=h433bba3_0
glib=2.69.1=h4ff587b_1
gputil=1.4.0=pypi_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py36h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=3.3=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
jasper=2.0.14=hd8c5072_2
jpeg=9e=h7f8727e_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1234567_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.1.0=2
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.16.4=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
opencv-python=4.6.0.66=pypi_0
openjpeg=2.4.0=h3ad879b_0
openslide=3.4.1=h8137273_1
openslide-python=1.1.2=pypi_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pypi_0
pcre=8.45=h295c915_0
pillow=6.1.0=pypi_0
pip=21.2.2=py36h06a4308_0
pixman=0.40.0=h7f8727e_1
py-opencv=3.4.2=py36hb342d67_1
pyparsing=3.0.9=pypi_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.1=py3.6_cuda10.1_cudnn7.6.3_0
pyzmq=23.2.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
scipy=1.5.2=py36h0b6359f_0
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h5082296_0
tk=8.6.12=h1ccaba5_0
torch=1.9.0+cu111=pypi_0
torchaudio=0.9.0=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.10.0+cu111=pypi_0
tornado=6.1=py36h27cfd23_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.12=pypi_0
visdom=0.1.8=pypi_0
websocket-client=1.3.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
